move AsyncStore tests to separate file
